messages don't work when reading from json

+++ server.py
@@ -2,6 +2,7 @@ import socket
 import sys
 import json
 import threading
+import unicodedata
 
 def processEntry(data, addr):
     user_exists = 0
@@ -31,7 +32,7 @@ def processEntry(data, addr):
                 json.dump(d_json, f)
                 f.write(']}')
     #  TODO update each users must recent action
-    else:
+    elif d_json['action'] != 'f':
         with open('example.json', 'a+') as f:
             f.seek(0, 2)
             f.truncate()
@@ -43,15 +44,14 @@ def processEntry(data, addr):
         solve_action(d_json)
 
 
-
-def give_users():
-    print("*GIVE USERS")
-    user_array = []
+def give_users(d_json):  # gives list of users connected to the chat
+    user_array = " "
     data = json.load(open('example.json'))
     for d in data['messages']:
-        user_array = d['user']
-        print("user_array", user_array)
-    #s.sendto(user_array, d['id'])
+        user = d['user'] + ", "
+        user_array += user
+    print("give_users", user_array, d_json['id'])
+    s.sendto(user_array, d_json['id'])
 
 
 def broadcast_message(d_json):
@@ -59,12 +59,13 @@ def broadcast_message(d_json):
     data = json.load(open('example.json'))
     for d in data['messages']:
         print("send to", d_json['text'], d['id'])
-        #s.sendto(d_json['text'], d['user'])
+        s.sendto(d_json['text'], d['user'])
 
 
 def private_message(d_json):
     user_exists = 0
     user_ok = 0
+    count = 0
     error_mssg = "User doesn't exist"
     text = d_json['text']
     array_text = text.split(',')
@@ -72,14 +73,21 @@ def private_message(d_json):
     recipient_pm = array_text[1]
     data = json.load(open('example.json'))
     for d in data['messages']:
+        count += 1
         if recipient_pm == d['user']:
-            user_exists = 1
+            tup = d['id']
+            modify_ip = tup[0]
+            modify_ip = modify_ip.encode("utf-8")
+            private_message = private_message.encode("utf-8")
+            tup[0] = modify_ip
+            # print("Private message sent: ", private_message, tuple(d['id']))
+            # print("Private message sent: ", private_message, tuple(tup))
+            print("Private message sent: ", private_message, d_json['id'])
+            s.sendto(private_message, tuple(tup))
         else:
-            user_ok = 1
-    if user_exists == 1:
-        print("Private message sent: ", private_message, tuple(d['id']))
-        s.sendto(private_message, tuple(d['id']))
-    elif user_ok == 1:
+            user_ok += 1
+    if user_ok == count:
+        print("error")
         s.sendto(error_mssg, d_json['id'])
 
 def exit(d_json):
@@ -93,13 +101,16 @@ def solve_action(d_json):
     if action == 'a':
         broadcast_message(d_json)
     elif action == 'b':
-        give_users()
+        give_users(d_json)
     elif action == 'c':
         private_message(d_json)
     elif action == 'd':
         exit(d_json)
     elif action == 'e':
         print("new user saved")
+    elif action == 'f':
+        print("reading")
+
 
 
 # Datagram (udp) socket
@@ -121,17 +132,19 @@ def server_thread():
 
         reply = 'OK...' + data
         processEntry(data, addr)
-        s.sendto(reply, addr)
-        print 'Message[' + addr[0] + ':' + str(addr[1]) + '] - ' + data.strip()
+        # s.sendto(reply, addr)
+        # print 'Message[' + addr[0] + ':' + str(addr[1]) + '] - ' + data.strip()
 
     s.close()
 
 
-HOST = '127.0.0.1'  # Symbolic name meaning all available interfaces
+HOST = ''  # Symbolic name meaning all available interfaces
 PORT = 8888  # Arbitrary non-privileged port
 
 try:
     s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)  #  Make Socket Reusable
+    s.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1) #  Allow incoming broadcasts
     print 'Socket created'
 except socket.error, msg:
     print 'Failed to create socket. Error Code : ' + str(msg[0]) + ' Message ' + msg[1]
